🚚 Moving files around
The target being to be closer to a 1 function per file so that we simplify tree-shakeability of the package and make users able to only pull the files they use and not all of them.
