+++ src/distribution/UnsafeUniformArrayIntDistribution.ts
@@ -1,4 +1,4 @@
-import { RandomGenerator } from '../generator/RandomGenerator';
+import { RandomGenerator } from '../types/RandomGenerator';
 import {
   addArrayIntToNew,
   addOneToPositiveArrayInt,
