+++ src/pure-rand-default.ts
@@ -1,16 +1,20 @@
-import { RandomGenerator, generateN, skipN, unsafeGenerateN, unsafeSkipN } from './generator/RandomGenerator';
+import { RandomGenerator } from './types/RandomGenerator';
 import { congruential32 } from './generator/LinearCongruential';
 import mersenne from './generator/MersenneTwister';
 import { xorshift128plus } from './generator/XorShift';
 import { xoroshiro128plus } from './generator/XoroShiro';
 
-import { Distribution } from './distribution/Distribution';
+import { Distribution } from './types/Distribution';
 import { uniformArrayIntDistribution } from './distribution/UniformArrayIntDistribution';
 import { uniformBigIntDistribution } from './distribution/UniformBigIntDistribution';
 import { uniformIntDistribution } from './distribution/UniformIntDistribution';
 import { unsafeUniformArrayIntDistribution } from './distribution/UnsafeUniformArrayIntDistribution';
 import { unsafeUniformBigIntDistribution } from './distribution/UnsafeUniformBigIntDistribution';
 import { unsafeUniformIntDistribution } from './distribution/UnsafeUniformIntDistribution';
+import { skipN } from './distribution/SkipN';
+import { generateN } from './distribution/GenerateN';
+import { unsafeGenerateN } from './distribution/UnsafeGenerateN';
+import { unsafeSkipN } from './distribution/UnsafeSkipN';
 
 // Explicit cast into string to avoid to have __type: "__PACKAGE_TYPE__"
 const __type = '__PACKAGE_TYPE__' as string;
